backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic[email]>=2.5.0
httpx>=0.25.0
groq>=0.9.0
dnspython>=2.4.0
